--- src/main/java/data/DukeException.java
@@ -1,12 +1,0 @@
-package data;
-
-public class DukeException extends Exception {
-    public DukeException(String errMsg) {
-        super(errMsg);
-    }
-
-    @Override
-    public String toString() {
-        return super.getMessage();
-    }
-}
