+++ src/main/java/duke/data/DukeException.java
@@ -1,0 +1,12 @@
+package duke.data;
+
+public class DukeException extends Exception {
+    public DukeException(String errMsg) {
+        super(errMsg);
+    }
+
+    @Override
+    public String toString() {
+        return super.getMessage();
+    }
+}
