+++ src/main/java/duke/task/Type.java
@@ -1,0 +1,5 @@
+package duke.task;
+
+public enum Type {
+    T, D, E
+}
