--- src/main/java/task/Type.java
@@ -1,5 +1,0 @@
-package task;
-
-public enum Type {
-    T, D, E
-}
