+++ text-ui-test/EXPECTED.TXT
@@ -1,38 +1,38 @@
     (/ /^ / - / ^/ /)
     ____________________________________________________________
-    Hello! I'm Duke by A0221330A.
+    Hello! I'm duke.Duke by A0221330A.
     What can I do for you?
     ____________________________________________________________
     ____________________________________________________________
-    Got it. I've added this task: 
+    Got it. I've added this duke.task:
         [T][ ] read book
     Now you have 1 tasks in the list.
     ____________________________________________________________
     ____________________________________________________________
-    Got it. I've added this task: 
+    Got it. I've added this duke.task:
         [D][ ] return book (by: Dec 20 2022)
     Now you have 2 tasks in the list.
     ____________________________________________________________
     ____________________________________________________________
-    Got it. I've added this task: 
+    Got it. I've added this duke.task:
         [E][ ] project meeting (at: Dec 21 2022)
     Now you have 3 tasks in the list.
     ____________________________________________________________
     ____________________________________________________________
-    Got it. I've added this task: 
+    Got it. I've added this duke.task:
         [T][ ] join sports club
     Now you have 4 tasks in the list.
     ____________________________________________________________
     ____________________________________________________________
-    Nice! I've marked this task as done:
+    Nice! I've marked this duke.task as done:
         [T][✓] read book
     ____________________________________________________________
     ____________________________________________________________
-    Nice! I've marked this task as done:
+    Nice! I've marked this duke.task as done:
         [T][✓] join sports club
     ____________________________________________________________
     ____________________________________________________________
-    Got it. I've added this task: 
+    Got it. I've added this duke.task:
         [T][ ] borrow book
     Now you have 5 tasks in the list.
     ____________________________________________________________
